后台用户增删改查关联用户角色权限

+++ app/Http/Controllers/Admin/PermissionController.php
@@ -9,15 +9,18 @@ use Spatie\Permission\Models\Permission;
 class PermissionController extends BaseController
 {
     public function index(){
+        $this->authorize('permission_index');
         $permissions=Permission::paginate(10);
         return view('admin.permission.index', compact('permissions'));
     }
 
     public function create(){
+        $this->authorize('permission_add');
         return view('admin.permission.create');
     }
 
     public function store(Request $request, Permission $permission){
+        $this->authorize('permission_add');
         $permission->name=$request->name;
         $permission->title=$request->title;
         $permission->save();
@@ -25,10 +28,12 @@ class PermissionController extends BaseController
     }
 
     public function edit(Permission $permission){
+        $this->authorize('permission_edit');
         return view('admin.permission.create', compact('permission'));
     }
 
     public function update(Permission $permission, Request $request){
+        $this->authorize('permission_edit');
         $permission->name=$request->name;
         $permission->title=$request->title;
         $permission->save();
@@ -36,6 +41,7 @@ class PermissionController extends BaseController
     }
 
     public function destroy(Permission $permission){
+        $this->authorize('permission_del');
         $permission->delete();
         return redirect()->route('permissions.index')->with('success','删除权限成功');
     }
